Update NFCRecord with proper type for data payload
NFC payloads always use data encoded as 8 bit, usually utf-8 strings.

Ideally we could return a Uint8Array for the data payload, but that's
not possible in the Capacitor bridge API. We only have access to the
types defined by Capacitor, which are the primitives like Array, Object
and number / boolean / string.

Now that the type is properly defined, it's more obvious that consumers
are responsible for coercing to a Uint8Array for decoding.