Preprocess `nfcTagRead` event so that data payload is always a DataView
The last release had an inconsistent API, where Web would return a
DataView with 8-bit values, while Android would return an Array<number>.

Cross-platform codebases would have to handle this in their
implementation by checking the type at runtime and converting if
necessary.

Now we always return a DataView, and so the consumer is setup for
success with the 8-bit payload that can be converted to utf-8 or
whatever their use case requires.